requirements.txt
Flask==3.0.3
PyJWT==2.9.0
python-dotenv==1.0.1
